reordering files (different folder for each different changing test)

+++ lib/no_weight_storage_modifiable_heap_random_selector.hpp
@@ -19,7 +19,7 @@ namespace stochastic {
 // }
 
   template <
-    typename I = int, size_t precision = std::numeric_limits<Real>::digits
+    typename I = size_t, size_t precision = std::numeric_limits<Real>::digits
   >
   class low_storage_selector :
     //Extends a complete tree...
